requirements.txt
flask_sessionstore
numpy
pillow
Flask==1.0.2
requests==2.20.1
uWSGI==2.0.18